VTTrac

Documentation for VTTrac.

VTTrac.chk_zmiss_region

VTTrac.chk_zsub_peak_inside

VTTrac.do_tracking

VTTrac.find_score_peak

VTTrac.find_subgrid_peak_5pt_epara

VTTrac.find_subgrid_peak_5pt_gaus

VTTrac.get_score

VTTrac.get_score_ncov

VTTrac.get_score_ncov_with_visible

VTTrac.get_score_with_visible

VTTrac.get_score_xcor

VTTrac.get_score_xcor_with_visible

VTTrac.get_zsub

VTTrac.get_zsub_subgrid

VTTrac.get_zsub_view

VTTrac.get_zsub_visible

VTTrac.get_zsub_visible_subgrid

VTTrac.get_zsub_visible_view

VTTrac.inspect_t_index

VTTrac.print_ary2d

VTTrac.roundInt

VTTrac.set_basic!

VTTrac.set_ixyhw_directly

VTTrac.set_ixyhw_from_v

VTTrac.set_optional!

VTTrac.setup

VTTrac.sliding_ncov

VTTrac.sliding_xcor

VTTrac.trac

VTTrac.chk_zmiss_region — Method

chk_zmiss_region(o, tid, k0, k1, l0, l1)

Check if there is data missing in the specified region at tid.

Returns

stat::Bool: false if there is no data missing, true if not.

VTTrac.chk_zsub_peak_inside — Method

chk_zsub_peak_inside(o, zs)

Check whether the template subimage is peaked (maximized or minimized) inside, and the peak is conspicuous enough, having a difference from the max or min on the sides greater than peakinsideth*(insidemax - insidemin).

Caution

If o.peak_inside_th < 0, no checking is conducted.

Returns

stat::Bool: false if passed the check, true if not.

VTTrac.do_tracking — Method

do_tracking(o, tid0, x0, y0, vx0, vy0, out_subimage, out_score_ary)

Conduct tracking (core).

Arguments

o::VTT: The traking oject.

tid0::Vector{Integer}: Tracking initial time indices.

x0::Vector{Float64}: Tracking initial template-center x location (index-based; non-integer for subgrid).

y0::Vector{Float64}: Tracking initial template-center y location (index-based; non-integer for subgrid).

vx0g::Vector{Float64}: First guess of vx (to search around it). Can be 0.

vy0g::Vector{Float64}: First guess of vy (to search around it). Can be 0.

out_subimage::Bool: Whether output subimages.

out_score_ary::Bool: Whether output score arrays.

Returns

count::Matrix{Float64}: (len: len) The number of successful tracking for each initial template.

tid::Matrix{Float64}: (len: (ntrac+1)*len) Time index of the trajectories (tid0 and subsequent ones).

x::Matrix{Float64}: (len: (ntrac+1)*len) x locations of the trajectories (x0 and derived ones).

y::Matrix{Float64}: (len: (ntrac+1)*len) y locations of trajectories (x0 and derived ones).

vx::Matrix{Float64}: (len: ntrac*len) Derived x-velocity.

vy::Matrix{Float64}: (len: ntrac*len) Derived y-velocity.

score::Matrix{Float64}: (len: ntrac*len) Scores along the trajectory (max values, possibly at subgrid).

zss::Array{Float32,4}: (optional, if non-nothing) (Diagnosis output if wanted) The subimages along the track (1D pointer for 4D array; nsx * nsy * (ntrac+1) * len.

score_ary::Array{Float32,4}: (optional, if non-nothing) (Diagnosis output if wanted) the entire scores (1D pointer for 4D array; (x-sliding size) * (y-sliding size) * (ntrac+1) * len.

VTTrac.find_score_peak — Method

find_score_peak(o, scr, kw, lw)

Find the score peak and its location.

Returns

stat::Bool: false if the peak is inside; true if not.

kpi::Float64: the peak location x.

lpi::Float64: the peak location y.

scrp::Float64: the peak score.

VTTrac.find_subgrid_peak_5pt_epara — Method

find_subgrid_peak_5pt_epara(c, l, r, b, t)

Find subgrid peak from 5 points with elliptic paraboloid.

Input c(enter), l(eft), r(ight), b(ottom), t(op) at (0,0), (-1,0), (0,1), (0,-1), (0,1), respectively. c must be greater than any of l,r,b,t.

Equation: z = -p(x-x0)^2 + -q(y-y0)^2 + r = -p x^2 + 2p x0 x - q y^2 + 2q y0 y + c

, where u = p x0, v = q y0, c = -p x0^2 - q y0^2 + r

Returns

stat::Bool: false if find peak successfully, true if not.

x0::float: x of peak the location.

y0::float: y of peak the location.

VTTrac.find_subgrid_peak_5pt_gaus — Method

find_subgrid_peak_5pt_gaus(c, l, r, b, t)

Find subgrid peak from 5 points by interpolating with a 2D gaussian (for positive scores).

It is simply a log-version of the elliptic-paraboloid method (find_subgrid_peak_5pt_epara). It appears that this method is preferred in many PIVs over the elliptic-paraboloid method (find_subgrid_peak_5pt_epara). In a test, there were not much difference between them, though.

Input c(enter), l(eft), r(ight), b(ottom), t(op) at (0,0), (-1,0), (0,1), (0,-1), (0,1), respectively. all of them must be positive. c must be greater than any of l,r,b,t.

VTTrac.get_score — Method

Conduct template matching driver

VTTrac.get_score_ncov — Method

get_score_ncov(o, x, tid, k0, k1, l0, l1)

Conduct template matching, scoring by normalized covariance.

Returns

stat::Bool: false if passed the check, true if not.

VTTrac.get_score_ncov_with_visible — Method

get_score_ncov_with_visible(o, x, tid, k0, k1, l0, l1)

Conduct template matching, scoring by normalized covariance.

Returns

stat::Bool: false if passed the check, true if not.

VTTrac.get_score_with_visible — Method

Conduct template matching driver with visible

VTTrac.get_score_xcor — Method

get_score_xcor(o, x, tid, k0, k1, l0, l1)

Conduct template matching, scoring by cross-correlation.

Returns

stat::Bool: false if passed the check, true if not.

VTTrac.get_score_xcor_with_visible — Method

get_score_xcor_with_visible(o, x, tid, k0, k1, l0, l1)

Conduct template matching, scoring by cross-correlation.

Returns

stat::Bool: false if passed the check, true if not.

VTTrac.get_zsub — Method

get_zsub(o, tid, xi, yi)

Read out a template subimage from the image at tid.

The sub-image positons are specified at its center. (If the sub-image size is even, with one more pix on the "left" / "bottom", by starting from the index xi-nsx/2, yi-nsy/2).

Returns

stats::Bool: false if successful (specified region is valid and, if chk_zmiss,

no data missing), true if not.

zsub::Matrix{Float32}: Subimage at (x,y) = (xi, yi).

See Also

get_zsub_view

VTTrac.get_zsub_subgrid — Method

get_zsub_subgrid(o, tid, x, y)

Read out a template submimage from the image at tid. Possibly at subgrid: Linearly interpolated, if x or y has deviation from integer (bilinear if x and y). Efficient: no unnecessary read-out is made.

Returns

stats::Bool: false if successful (specified region is valid and, if chk_zmiss,

no data missing), true if not.

zsubg::Matrix{Float32}: Subimage.

VTTrac.get_zsub_view — Method

get_zsub_view(o, tid, xi, yi)

Like get_zsub, but returns a view.

Notes

This function returns a view of subimage. get_zsub returns a copy of subimage.

See Also

get_zsub

VTTrac.get_zsub_visible — Method

get_zsub_visible(o, tid, xi, yi)

Read out a template subimage and subvisible from the image and visible at tid.

The sub-image positons are specified at its center. (If the sub-image size is even, with one more pix on the "left" / "bottom", by starting from the index xi-nsx/2, yi-nsy/2).

Returns

stats::Bool: false if successful (specified region is valid and, if chk_zmiss,

no data missing), true if not.

zsub::Matrix{Float32}: Subimage at (x,y) = (xi, yi).

See Also

get_zsub_visible_view

VTTrac.get_zsub_visible_subgrid — Method

get_zsub_visible_subgrid(o, tid, x, y)

Read out a template submimage from the image at tid. Possibly at subgrid: Linearly interpolated, if x or y has deviation from integer (bilinear if x and y). Efficient: no unnecessary read-out is made.

Returns

stats::Bool: false if successful (specified region is valid and, if chk_zmiss,

no data missing), true if not.

zsubg::Matrix{Float32}: Subimage.

VTTrac.get_zsub_visible_view — Method

get_zsub_visible_view(o, tid, xi, yi)

Like get_zsub_visible, but returns a view.

Notes

This function returns a view of zsub and visible. get_zsub_visible returns a copy of zsub and visible.

See Also

get_zsub_visible

VTTrac.inspect_t_index — Method

To check whether a time index is valid. Returns false if valid, true if not.

VTTrac.print_ary2d — Method

print a 2D double array

VTTrac.roundInt — Method

round to Int like C/C++

VTTrac.set_basic! — Method

set_basic!(o, nsx, nsy, itstep, ntrac)

Sets basic (mandatory) tracking parameters. Also sets default vals for optional params.

Arguments

o::VTT: The object.

nsx::Integer: The template subimage size (x).

nsy::Integer: The template subimage size (y).

itstep::Integer: Index-based time step for tracking (can be negative).

ntrac::Integer: Number of times for each initial template is tracked.

VTTrac.set_ixyhw_directly — Method

set_ixyhw_from_v(o, ixch, iyxh)

Sets the tracking parameters i[xy]hw.

Arguments

o::VTT: The object.

ixhw::Float64: The range over which next x is searched around initial guess.

iyhw::Float64: The range over which next y is searched around initial guess.

VTTrac.set_ixyhw_from_v — Method

set_ixyhw_from_v(o, vxch, vyxh)

Sets the tracking parameters i[xy]hw from velocities (v[xy]hh).

Arguments

o::VTT: The object.

vxhw::Float64: The range over which vx is searched around initial guess.

vyhw::Float64: The range over which vy is searched around initial guess.

```

VTTrac.set_optional! — Method

set_optional!(o, subgrid, subgrid_gaus, score_method, score_th0, score_th1, peak_inside_th, min_contrast, vxch, vych, use_init_temp)

Sets optional tracking parameters.

Arguments

o::VTT: The object.

subgrid::Bool: Whether to conduct subgrid tracking.

subgrid_gaus::Bool: Whether subgrid peak finding is by gaussian.

score_method::String: Scoring method (such as xcor for cross-correlation).

score_th0::AbstractFloat: (Result screening parameter) Minimum score required for the first-time tracking.

score_th1::AbstractFloat: (Result screening parameter) Minimum score required for the subsequent tracking.

peak_inside_th::Real: An initial template is used only when it is peaked (max or min) inside, exceeding the max or min along the sides by the ratio specified by its value.

min_contrast::Real: An initial template is used only when it has a difference in max and min greater than its value.

vxch::Float64: (Result screening parameter) If positive, tracking result is rejected if the vx chnages along trajecty greather than this value (thus used only when ntrac>=2). As a special case, if the result of the second tracking is rejected, the first one is also rejected, since there is no consecutive consistent result in this case.

vych::Float64: (Result screening parameter) As vxch but for the y-component.

min_visible::Int: Minimum number of visible values to calculate score when chk_mask is true.

VTTrac.setup — Method

setup(o, nsx, nsy, vxhw, vyhw, [ixhw, iyhw, subgrid, subgrid_gaus,
    itstep, ntrac, score_method, score_th0, score_th1, vxch, vych,
    peak_inside, peak_inside_th, min_contrast, use_init_temp, min_visible])

Setup for tracking.

Arguments

o::VTT: The object.

nsx::Integer, nsy::Integer: Submimage x & y sizes (x:1st, y:2nd dim).

vxch::Union{Real, Nothing}, vyhw::Union{Real, Nothing}: (either v[xy]hw or i[xy]hw are MANDATORY). the dimensions along which to perform the computation. search velocity range half sizes to set i[xy]hw. Seach at least to cover +-v?hw around the first guess or previous step. (the result can be outside the range.)

ixhw::Union{Int, Nothing}, iyhw::Union{Int, Nothing}: (either v[xy]hw or i[xy]hw are MANDATORY) Max displacement fro template match (can be set indirecly through v[xy]hw).

subgrid::Bool=true: Whether to conduct subgrid tracking.

subgrid_gaus::Bool=true: Whether subgrid peak finding is by gaussian.

itstep::Integer=1: Step of t's used (skip if >1).

ntrack::Integer=2: Max tracking times from initial loc.

score_method::String="xcor": "xcor" for cross-correlation, "ncov" for normalized covariance.

score_th0::AbstractFloat=0.8: The minimum score required for the 1st tracking.

score_th1::AbstractFloat=0.7: The minimum score required for subsequent tracking.

vxch::Union{Real, Nothing}=nothing: If non-nothing, the max tolerant vx change between two consecutive tracking.

vych::Union{Real, Nothing}=nothing: If non-nothing, the max tolerant vy change between two consecutive tracking.

peak_inside_th::Union{Real, Nothing}=nothing: If non-nothing, an initial template is used only when it is peaked (max or min) inside, exceeding the max or min along the sides by the ratio specified by its value.

min_contrast::Union{Real, Nothing}=nothing: If non-nothing, an initial template is used only when it has a difference in max and min greater than its value.

min_visible::Int=1: Minimum number of visible values to calculate score when chk_mask is true.

VTTrac.sliding_ncov — Method

sliding_ncov(o, sigx, xd, tid, k0, k1, l0, l1)

Sliding normalized covariance between the sugimage and image at tid.

Normalization is done by the sigma of the fist image : cov(x',y')/sigx^2 (in contrast to cov(x',y')/sigx/sigy in the correlation coefficient).

Returns

stat::Bool: false if all the relevant data and regions are valid, so all the scores (ncov) are defined at all tested center locations; true if not.

scr::Matrix{Float32}: Score array.

VTTrac.sliding_xcor — Method

sliding_xcor(o, sigx, xd, tid, k0, k1, l0, l1)

Sliding cross-correlation between the sugimage and image at tid.

Returns

stat::Bool: false if all the relevant data and regions are valid, so all the scores (xcor) are defined at all tested center locations; true if not.

scr::Matrix{Float32}: Score array.

VTTrac.trac — Method

trac(o, tid, x, y[, vxg, vyg, out_subimage, out_score_ary])

Conduct tracking.

Arguments

o::VTT: The traking oject.

tid::Array{Integer,Any}: Tracking initial time indices.

x::Array{Float64,Any}: Tracking initial template-center x location (index-based; non-integer for subgrid).

y::Array{Float64,Any}: Tracking initial template-center y location (index-based; non-integer for subgrid).

vxg::Array{Float64,Any}=nothing: First guess of vx (to search around it). Can be 0.

vyg::Array{Float64,Any}=nothing: First guess of vy (to search around it). Can be 0.

out_subimage::Bool=false: Whether output subimages.

out_score_ary::Bool=false: Whether output score arrays.

to_missing::Bool=true: Whether output missing values as missing.

Returns

count::Vector{Integer}: [len] The number of successful tracking for each initial template.

tid::Matrix{Float64}: [ntrac+1, len] time index of the trajectories (tid0 and subsequent ones).

x::Matrix{Float64}: [ntrac+1, len] x locations of the trajectories (x0 and derived ones).

y::Matrix{Float64}: [ntrac+1, len] y locations of trajectories (x0 and derived ones).

vx::Matrix{Float64}: [ntrac, len] Derived x-velocity.

vy::Matrix{Float64}: [ntrac, len] Derived y-velocity.

score::Matrix{Float64}: [ntrac, len] Scores along the trajectory (max values, possibly at subgrid).

zss::Array{Float32,4}: nsx, nsy, ntrac+1, len (Diagnosis output if wanted) The subimages along the track.

score_arry::Array{Float64,4}: (x-sliding size, y-sliding size, ntrac+1, len (Diagnosis output if wanted) The entire scores.