VTTrac.setup — Function

setup(o, nsx, nsy, vxhw, vyhw, [ixhw, iyhw, subgrid, subgrid_gaus,
    itstep, ntrac, score_method, score_th0, score_th1, vxch, vych,
    peak_inside, peak_inside_th, min_contrast, use_init_temp, min_visible])

Setup for tracking.

Arguments

o::VTT: The object.

nsx::Integer, nsy::Integer: Submimage x & y sizes (x:1st, y:2nd dim).

vxch::Union{Real, Nothing}, vyhw::Union{Real, Nothing}: (either v[xy]hw or i[xy]hw are MANDATORY). the dimensions along which to perform the computation. search velocity range half sizes to set i[xy]hw. Seach at least to cover +-v?hw around the first guess or previous step. (the result can be outside the range.)

ixhw::Union{Int, Nothing}, iyhw::Union{Int, Nothing}: (either v[xy]hw or i[xy]hw are MANDATORY) Max displacement fro template match (can be set indirecly through v[xy]hw).

subgrid::Bool=true: Whether to conduct subgrid tracking.

subgrid_gaus::Bool=true: Whether subgrid peak finding is by gaussian.

itstep::Integer=1: Step of t's used (skip if >1).

ntrack::Integer=2: Max tracking times from initial loc.

score_method::String="xcor": "xcor" for cross-correlation, "ncov" for normalized covariance.

score_th0::AbstractFloat=0.8: The minimum score required for the 1st tracking.

score_th1::AbstractFloat=0.7: The minimum score required for subsequent tracking.

vxch::Union{Real, Nothing}=nothing: If non-nothing, the max tolerant vx change between two consecutive tracking.

vych::Union{Real, Nothing}=nothing: If non-nothing, the max tolerant vy change between two consecutive tracking.

peak_inside_th::Union{Real, Nothing}=nothing: If non-nothing, an initial template is used only when it is peaked (max or min) inside, exceeding the max or min along the sides by the ratio specified by its value.

min_contrast::Union{Real, Nothing}=nothing: If non-nothing, an initial template is used only when it has a difference in max and min greater than its value.

min_visible::Int=1: Minimum number of visible values to calculate score when chk_mask is true.

VTTrac.set_ixyhw_from_v — Function

set_ixyhw_from_v(o, vxch, vyxh)

Sets the tracking parameters i[xy]hw from velocities (v[xy]hh).

Arguments

o::VTT: The object.

vxhw::Float64: The range over which vx is searched around initial guess.

vyhw::Float64: The range over which vy is searched around initial guess.

```

VTTrac.set_ixyhw_directly — Function

set_ixyhw_from_v(o, ixch, iyxh)

Sets the tracking parameters i[xy]hw.

Arguments

o::VTT: The object.

ixhw::Float64: The range over which next x is searched around initial guess.

iyhw::Float64: The range over which next y is searched around initial guess.

VTTrac.trac — Function

trac(o, tid, x, y[, vxg, vyg, out_subimage, out_score_ary])

Conduct tracking.

Arguments

o::VTT: The traking oject.

tid::Array{Integer,Any}: Tracking initial time indices.

x::Array{Float64,Any}: Tracking initial template-center x location (index-based; non-integer for subgrid).

y::Array{Float64,Any}: Tracking initial template-center y location (index-based; non-integer for subgrid).

vxg::Array{Float64,Any}=nothing: First guess of vx (to search around it). Can be 0.

vyg::Array{Float64,Any}=nothing: First guess of vy (to search around it). Can be 0.

out_subimage::Bool=false: Whether output subimages.

out_score_ary::Bool=false: Whether output score arrays.

to_missing::Bool=true: Whether output missing values as missing.

Returns

count::Vector{Integer}: [len] The number of successful tracking for each initial template.

tid::Matrix{Float64}: [ntrac+1, len] time index of the trajectories (tid0 and subsequent ones).

x::Matrix{Float64}: [ntrac+1, len] x locations of the trajectories (x0 and derived ones).

y::Matrix{Float64}: [ntrac+1, len] y locations of trajectories (x0 and derived ones).

vx::Matrix{Float64}: [ntrac, len] Derived x-velocity.

vy::Matrix{Float64}: [ntrac, len] Derived y-velocity.

score::Matrix{Float64}: [ntrac, len] Scores along the trajectory (max values, possibly at subgrid).

zss::Array{Float32,4}: [nsx, nsy, ntrac+1, len] optional, if non-nothing (Diagnosis output if wanted) The subimages along the track.

score_arry::Array{Float64,4}: [(x-sliding size, y-sliding size, ntrac+1, len] optional, if non-nothing (Diagnosis output if wanted) The entire scores.